Added a bunch of core fish classes, reorganised the animal base classes, as well as made some edits to the variant context interface.
