+++ common/src/main/java/com/collective/projectcore/entities/variant/VariantContext.java
@@ -6,10 +6,27 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Random;
 
+/**
+ * An interface for the calculation and application of variants for entities that require them.
+ * First, instantiate this class in each creature class that requires it.
+ * Then, for each type of colour morph, create a record in your entity class which implements...
+ * @see VariantMorph
+ * Then fill in the details for the record by implementing the required methods.
+ * Use AmericanRedFoxEntity in Project Wildlife as an example.
+ */
 public interface VariantContext {
 
+    /**
+     * Contains all entity-specific morph records.
+     */
     List<VariantMorph> morphs();
+
+    /**
+     * Determines the rarity of wild-spawning variants.
+     * Will need to be defined when VariantContext is created.
+     */
     String calculateWildFunc();
+
     default String calculateVariant(String parent1, String parent2) {
         Random random = new Random();
         int parent1Light = 1;
@@ -26,10 +43,6 @@ public interface VariantContext {
         List<VariantMorph> possibleMorphs = new ArrayList<>();
         for (VariantMorph c : morphs()) {
             if (c.possibleLightLevels().contains(totalLightValue)) {
-                // Picks a morph based on...
-                // ...the rarity (defined below) +1 (to account for 0 rarities)...
-                // ...and then that times rarity + itself to create an exponential curve.
-                // Wanted to do rarity+1 * rarity*rarity, but I think the high numbers would be too resource intensive...
                 for (int i = 0; i <= (c.rarity()+1)*(c.rarity()+c.rarity()); i++) {
                     possibleMorphs.add(c);
                 }
@@ -38,6 +51,9 @@ public interface VariantContext {
         return possibleMorphs.get(random.nextInt(possibleMorphs.size())).name();
     }
 
+    /**
+     * Implemented by morph records created in specific entity classes to be used by <code>VariantContext</code>.
+     */
     interface VariantMorph {
         String name();
         int lightValue();
